+++ src/app/(home)/_services/books.ts
@@ -4,22 +4,23 @@ import { BookItemData, PopularBookData } from '@/types';
 const formatePopularBookDataToBookItemData = (doc: PopularBookData) => {
   let author = doc.authors;
   let translator;
+
   if (doc.authors.includes('지은이')) {
     author = doc.authors
       .split(';')[0]
       .trim()
-      .replace(/지은이|:/, '');
+      .replace(/지은이:/, '');
   }
   if (doc.authors.includes('옮긴이')) {
     translator = doc.authors
       .split(';')[1]
       .trim()
-      .replace(/옮긴이|:/, '');
+      .replace(/옮긴이:/, '');
   }
 
   const book: BookItemData = {
     title: doc.bookname,
-    isbn: doc.isbn13,
+    isbn: Number(doc.isbn13),
     author,
     translator,
     publisher: doc.publisher,
