+++ src/app/(home)/page.tsx
@@ -10,13 +10,13 @@ const Home = async () => {
   const hotBooks = BOOK_LIST_MOCK_DATA;
   const lastMonthLibrarianPick = await getLastMonthLibrarianPick();
   const popularBooks = await getPopularBooks();
-  console.log('✨✨✨✨✨', popularBooks);
 
   return (
     <div className={styles.homeContainer}>
       <h1 className="sr-only">BOOKER 홈페이지</h1>
       <BooksCarouselSection.Loaded title="사서 추천 도서" booksSimpleInfo={lastMonthLibrarianPick} />
       <BookListSection.Loaded title="인기 대출 도서" bookItemsData={popularBooks} />
+      <BookListSection.Loaded title="급상승! 대출 도서" bookItemsData={hotBooks} />
       <BookListSection.Loaded title="다독자를 위한 추천 도서" bookItemsData={hotBooks} />
     </div>
   );
