+++ src/mocks/handlers/books.ts
@@ -1,8 +1,10 @@
 import { http, HttpResponse } from 'msw';
 
 import { publicLibraryEndpoint } from '@/services';
+import { naruEndpoint } from '@/services/endpoints/naruEndpoint';
 
 import { LIBRARIAN_PICK_XML } from '../mockData';
+import { POPULAR_BOOKS_DATA } from '../mockData/books';
 
 const interceptGetLastMonthLibrarianPick = () => {
   return http.get(publicLibraryEndpoint.librarianPick, async () => {
@@ -10,4 +12,10 @@ const interceptGetLastMonthLibrarianPick = () => {
   });
 };
 
-export const booksHandlers = [interceptGetLastMonthLibrarianPick()];
+const interceptGetPopularBooks = () => {
+  return http.get(naruEndpoint.popularBooks, async () => {
+    return HttpResponse.json(POPULAR_BOOKS_DATA);
+  });
+};
+
+export const booksHandlers = [interceptGetLastMonthLibrarianPick(), interceptGetPopularBooks()];
